Moved core logic to app.
