+++ Shivers Randomizer_x64/MEMORY_BASIC_INFORMATION64.cs
@@ -1,0 +1,21 @@
+﻿using System.Runtime.InteropServices;
+
+namespace Shivers_Randomizer_x64
+{
+    public partial class AttachPopup_x64
+    {
+        [StructLayout(LayoutKind.Sequential)]
+        public struct MEMORY_BASIC_INFORMATION64
+        {
+            public ulong BaseAddress;
+            public ulong AllocationBase;
+            public int AllocationProtect;
+            public int __alignment1;
+            public ulong RegionSize;
+            public int State;
+            public int Protect;
+            public int Type;
+            public int __alignment2;
+        }
+    }
+}
