+++ Shivers Randomizer_x64/MainWindow_x64.xaml
@@ -3,7 +3,6 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:Shivers_Randomizer_x64"
         mc:Ignorable="d"
         Title="Shivers Randomizer" Height="800" Width="355.738" Loaded="Window_Loaded">
     <Grid>
