+++ Shivers Randomizer_x64/Overlay_x64.xaml
@@ -7,6 +7,6 @@
         mc:Ignorable="d"
         Title="Overlay" Height="300" Width="400" AllowsTransparency="true" WindowStyle="None" Background="Transparent" Topmost="True" ShowInTaskbar="False">
     <Canvas>
-        <Label Name="labelOverlay" Content="Test Overlay Text"  HorizontalAlignment="Left" VerticalAlignment="Top" Foreground="Lime" Canvas.Left="10" Canvas.Top="24"/>
+        <Label Name="labelOverlay" Content="Test Overlay Text"  HorizontalAlignment="Left" VerticalAlignment="Center" Foreground="Lime" Canvas.Left="10"/>
     </Canvas>
 </Window>
